Add a method for merging overlay data into style, but not implemented yet
